Ponytail refactor: fix blockers B1-B6 + production-ready local Docker
Blockers (from staff-engineer review):
- B1: PositionLimitValidator always returned :passed (both ternary branches identical).
- B2: MarginValidator qty.to_i zeroed every fractional crypto order's notional.
- B3: LiquidationJob deadlocked via submit_order — added `internal:` flag that
  skips the margin lock and risk gate for closing orders.
- B4: test-api-key-123 mapping in BaseController was active in production.
- B5: docker-compose.yml baked RAILS_MASTER_KEY/SECRET_KEY_BASE as plaintext defaults.
- B6: LiquidationEngine cache mutation not atomic — wrapped in mutex.synchronize.

High-severity:
- H1: submit_order wraps lock+risk+fill in one transaction.
- H4: OrderBook#default_book raises in production when no market data.
- H5: OrdersController rescues specific exceptions.
- H6: LatencyEngine.simulate off by default.
- M13: Controller permits `context: {}` so VixGateValidator fires.

Ponytail-review deletions (net -73 LOC): SettlementEngine (no-op),
tmp_boot_debug.rb, test/ minitest cruft, config/initializers/redis.rb,
config/initializers/services_loader.rb, MarketData::OrderEvent, unused
order_params, duplicate kebab-case routes.

Other: Account validator allows negative available_balance/current_equity
(legitimate liquidated state). OrderValidator raises OrderValidationError.
Dockerfile split into development + production stages. Compose uses
env_file: .env with healthcheck. Smoke-test uses tsx, snake_case routes.

Regression specs: B1 (over-limit reject), B2 (fractional-crypto reject),
B3 (liquidation-flow integration spec), OrderValidator ActionController::Parameters path.

+++ app/services/risk/margin_validator.rb
@@ -5,7 +5,7 @@ module Risk
 
     def evaluate(account_id, signal)
       symbol = signal.to_h[:symbol]
-      qty = signal.to_h[:quantity].to_i
+      qty = signal.to_h[:quantity].to_f   # was .to_i — silently zeroed every fractional crypto order
       price = (signal.to_h[:price] || signal.to_h[:ltp]).to_f
       instrument_type = signal.to_h[:instrument_type].to_s.presence || "EQUITY"
       notional = price * qty
